backend/requirements.txt
anyio==4.2.0
asgiref==3.7.2
asttokens==2.2.1
attrs==24.2.0
autopep8==2.0.2
backcall==0.2.0
beautifulsoup4==4.12.2
bitarray==2.7.6
bleach==6.2.0
bs4==0.0.1
certifi==2024.8.30
cffi==1.17.0
charset-normalizer==3.4.0
click==8.1.3
colorama==0.4.6
comm==0.1.2
contourpy==1.0.7
cycler==0.11.0
dash==2.17.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dataclasses-json==0.6.7
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
Django==5.0.1
docopt==0.6.2
et-xmlfile==1.1.0
executing==1.2.0
fastjsonschema==2.16.3
filelock==3.12.0
Flask==2.2.3
fonttools==4.39.3
geographiclib==2.0
geolocation-python==0.2.2
geopy==2.3.0
greenlet==3.0.3
gTTS==2.3.2
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.10
importlib_metadata==7.1.0
ipykernel==6.21.1
ipython==8.12.3
isodate==0.7.2
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jsonschema==4.17.3
jupyter_client==8.0.2
jupyter_core==5.2.0
jupyterlab_pygments==0.3.0
keyboard==0.13.5
kiwisolver==1.4.4
MarkupSafe==2.1.2
marshmallow==3.23.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==3.1.2
mpmath==1.3.0
multiset==3.0.2
mypy-extensions==1.0.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.8.0
nest-asyncio==1.5.6
numpy==1.24.2
oauthlib==3.2.2
opencv-python==4.7.0.72
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==24.1
pandas==2.0.0
pandocfilters==1.5.1
parso==0.8.3
pdfreader==0.1.12
pickleshare==0.7.5
Pillow==9.5.0
pipreqs==0.5.0
platformdirs==3.2.0
playsound==1.3.0
plotly==5.14.1
prompt-toolkit==3.0.36
psutil==5.9.4
psycopg2==2.9.9
psycopg2-binary==2.9.9
pure-eval==0.2.2
py4j==0.10.9.7
pyaes==1.6.1
pyasn1==0.5.0
pycodestyle==2.10.0
pycparser==2.22
pycryptodome==3.18.0
pygame==2.4.0
Pygments==2.14.0
pyparsing==3.0.9
pypdf==4.3.1
PyPDF2==3.0.1
pyrsistent==0.19.3
PySocks==1.7.1
pyspark==3.5.3
python-dateutil==2.8.2
python-dotenv==1.0.1
python-telegram-bot==20.7
python-youtube==0.9.7
pytz==2023.3
pywin32==305
pyzmq==25.0.0
requests==2.32.3
requests-oauthlib==1.3.0
retrying==1.3.4
rsa==4.9
scipy==1.12.0
selenium==4.25.0
shapely==2.0.3
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.4.1
SQLAlchemy==2.0.28
sqlparse==0.4.4
stack-data==0.6.2
sympy==1.12
Telethon==1.32.1
tenacity==8.2.2
tinycss2==1.4.0
tinydb==4.8.0
tornado==6.2
traitlets==5.9.0
trio==0.27.0
trio-websocket==0.11.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2023.3
urllib3==2.2.3
virtualenv==20.22.0
waitress==2.1.2
wcwidth==0.2.6
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==2.2.3
wsproto==1.2.0
yapf==0.32.0
yarg==0.1.9
youtube-transcript-api==0.6.3
zipp==3.18.2
